ParEval/requirements.txt
datasets==3.3.2
google-generativeai==0.8.5
huggingface-hub==0.29.2
multiprocess==0.70.15
numpy>=2.0.1
pandas>=2.2.3
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.49.0
alive-progress>=2.0.0
vllm
flashinfer-python
sacrebleu
rouge_score
codebleu
tree-sitter-python==0.21
evaluate